requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2
django==4.1
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.2.1
sqlparse==0.4.2
gunicorn
